bump dependencies inplace dependency
