+++ server/.env
@@ -1,0 +1,1 @@
+SECRET=dwaf0j4f43jf93f0349f3f093jff
