--- server/config.json
@@ -1,3 +1,0 @@
-{
-  "token": "b84561a85321c36982e18bfcdfc081bce8f7ca9b"
-}
