+++ server/src/index.ts
@@ -1,3 +1,5 @@
+import 'reflect-metadata'
+
 import {app} from './app'
 import {createConnection} from 'typeorm'
