+++ server/src/routes/participation.ts
@@ -1,7 +1,7 @@
 import * as Router from 'koa-router'
 import {participationController} from '../controllers/participation'
-import validate = require('koa-joi-validate')
 import {participationValidator} from '../validators/participation'
+import {validateMiddleware} from '../helpers'
 
 export const participationRouter = new Router()
     .prefix('/participation')
@@ -12,14 +12,14 @@ export const participationRouter = new Router()
         return next()
     })
     .post('/placement/:placementId',
-        validate(participationValidator.create),
+        validateMiddleware(participationValidator.create),
         async ctx => {
             const placementId = parseInt(ctx.params.placementId, 10)
 
             await participationController.create(ctx, placementId, ctx.state.user)
         })
     .delete('/:id',
-        validate(participationValidator.delete),
+        validateMiddleware(participationValidator.delete),
         async ctx => {
             const participationRequestId = parseInt(ctx.params.id, 10)
