+++ server/tsconfig.json
@@ -1,19 +1,16 @@
 {
   "compilerOptions": {
-    "target": "es2017",
-    "lib": [
-      "esnext"
-    ],
+    "allowJs": true,
     "module": "commonjs",
-    "removeComments": true,
-    "rootDir": "src",
-    "outDir": "build",
-    "emitDecoratorMetadata": true,
+    "target": "es2017",
+    "lib": ["esnext"],
+    "moduleResolution": "node",
     "experimentalDecorators": true,
-    "strict": true,
-    "strictPropertyInitialization": false
+    "emitDecoratorMetadata": true,
+    "sourceMap": true,
+    "outDir": "dist"
   },
-  "exclude": [
-    "node_modules"
+  "include": [
+    "src/**/*"
   ]
 }
