+++ server/tslint.json
@@ -28,7 +28,7 @@
     "space-before-function-paren": false,
     "no-var-requires": false,
     "no-console": false,
-    "no-implicit-dependencies": ["dev"],
+    "no-implicit-dependencies": [true, "dev"],
     "no-empty-interface": false
   },
   "rulesDirectory": []
